Download satellite ocean colour data

Downloads satellite ocean colour bloom metrics generated by the SOPhyE team at DFO from the CIOOS Atlantic ERDDAP (see "Details" for coordinate boundaries and years for which data is available). Can be used to download other datasets if the dataset ID and ERDDAP URL is specified.

Usage

get_erddap_data(
  dataset_id = "bio_remote_sensing_occci_nwa_poly4_spring_bloom",
  erddap_url = "https://cioosatlantic.ca/erddap/",
  variables = c("t_start", "t_duration", "amplitude_real", "magnitude_real",
    "annual_mean", "NRMSE_bloom", "percent_dineof"),
  xlim = NULL,
  ylim = NULL,
  timelim = NULL
)

Arguments

dataset_id

Dataset ID (default is "bio_remote_sensing_occci_nwa_poly4_spring_bloom", the satellite ocean colour spring bloom metrics generated by the SOPhyE group at DFO)

erddap_url

URL for the ERDDAP server where the data is located (default is "https://cioosatlantic.ca/erddap/", the CIOOS Atlantic ERDDAP server)

variables

vector of variable names to download (default is the vector of variables in the spring bloom metrics dataset)

xlim

Minimum and maximum longitude (if NULL, downloads the full range of longitudes)

ylim

Minimum and maximum latitude (if NULL, downloads the full range of latitudes)

timelim

Start and end datetime string in format "YYYY-MM-DDT00:00:00Z" (if NULL, downloads the full time series)

Value

Dataframe with columns longitude, latitude, time, and the selected variables

Details

SOPhyE satellite ocean colour spring bloom metrics: 42-76 degrees west, 39-66 degrees north 1998-2024 (will be updated annually) Note: Years should be formatted as e.g. "1998-01-01T00:00:00Z" in the "timelim" argument

The satellite bloom metrics downloaded by this function are 4.64 km resolution and include all the variables in the product at a lower level of quality control. In contrast, the satellite_ocean_colour variable available in the marea package is fully quality-controlled and ready to use, excludes the two quality indicator layers (NRMSE_bloom and percent_dineof), and has been resampled to 0.1 degree resolution (it is recommended to use the mean or median of pixel values within this range when working with the original data, so this is the appropriate resolution to use).

Examples

if (FALSE) { # \dontrun{
df <- get_erddap_data(
  variables = c("t_start", "t_duration"),
  timelim = c("2023-01-01T00:00:00Z", "2024-01-01T00:00:00Z")
)
r <- stars::st_as_stars(df, dims = c("longitude", "latitude", "time"))
plot(r)
} # }